pyproject.toml
[tool.poetry]
name = "diary-project"
version = "0.1.0"
description = "Django diary application"
authors = ["Developer <dev@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
django = ">=4.2.0,<5.0.0"
psycopg2-binary = ">=2.9.10,<3.0.0"
python-dotenv = ">=1.1.0,<2.0.0"
django-environ = ">=0.12.0,<0.13.0"
gunicorn = ">=23.0.0,<24.0.0"
whitenoise = ">=6.9.0,<7.0.0"
django-debug-toolbar = ">=5.2.0,<6.0.0"
django-extensions = ">=4.1,<5.0"
django-seed = ">=0.3.1,<0.4.0"
pytest = ">=8.3.5,<9.0.0"
pytest-django = ">=4.11.1,<5.0.0"
pytest-describe = ">=2.2.0,<3.0.0"
faker = ">=37.3.0,<38.0.0"
ipython = "^9.2.0"
ipdb = "^0.13.13"
debugpy = "^1.8.14"
factory-boy = "^3.3.3"
playwright = "^1.40.0"
python3-saml = "^1.16.0"
django-guardian = "^3.0.0"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
python_classes = ["Test*", "*Test", "*Tests"]
python_functions = ["test_*"]
testpaths = ["apps"]
addopts = ["--reuse-db", "--nomigrations", "-v"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
